Return void from validator functions, not of bool
The validator functions used to return a boolean, which was true
if the check was valid, or threw an exception if the check failed.

This made returning any value pointless, as the successful
execution of the function denoted the check passed, and if it
failed, an exception was thrown.

So I've changed the return value from boolean to void.