PolicyEngine is abstract class, not interface
The Policy Engine can have multiple functions. It depends on how
the future implementers decide to use it. So it's best to leave it
as an abstract class with the methods I think they'll need.

This way, if they decide to use their own methods, they can easily
do it.